+++ sdk/assemblyscript/src/assembly/index.ts
@@ -22,9 +22,6 @@ export { graphql };
 import * as http from "./http";
 export { http };
 
-import * as collections from "./collections";
-export { collections };
-
 import models from "./models";
 export { models };